split up the available tests
