+++ examples/basic_routing/outputs.tf
