+++ examples/basic_routing/variables.tf
