+++ test/basic_routing_test.go
@@ -1,0 +1,22 @@
+package test
+
+import (
+	"github.com/gruntwork-io/terratest/modules/terraform"
+	"testing"
+)
+
+func TestCompleteExample(t *testing.T) {
+	t.Parallel()
+
+	terraformOptions := &terraform.Options{
+		// The path to where your Terraform code is located
+		TerraformDir: "../examples/basic_routing",
+		Upgrade:      true,
+	}
+
+	// At the end of the test, run `terraform destroy` to clean up any resources that were created
+	defer terraform.Destroy(t, terraformOptions)
+
+	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
+	terraform.InitAndApply(t, terraformOptions)
+}
