+++ test/private_hosted_zone_test.go
@@ -1,0 +1,22 @@
+package test
+
+import (
+	"github.com/gruntwork-io/terratest/modules/terraform"
+	"testing"
+)
+
+func TestPrivateHostedZone(t *testing.T) {
+	t.Parallel()
+
+	terraformOptions := &terraform.Options{
+		// The path to where your Terraform code is located
+		TerraformDir: "../examples/private-hosted-zone",
+		Upgrade:      true,
+	}
+
+	// At the end of the test, run `terraform destroy` to clean up any resources that were created
+	defer terraform.Destroy(t, terraformOptions)
+
+	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
+	terraform.InitAndApply(t, terraformOptions)
+}
