--- test/route53_test.go
@@ -1,86 +1,0 @@
-package test
-
-import (
-	"github.com/gruntwork-io/terratest/modules/terraform"
-	"testing"
-)
-
-func TestCompleteExample(t *testing.T) {
-	t.Parallel()
-
-	terraformOptions := &terraform.Options{
-		// The path to where your Terraform code is located
-		TerraformDir: "../examples/complete-example",
-		Upgrade:      true,
-	}
-
-	// At the end of the test, run `terraform destroy` to clean up any resources that were created
-	defer terraform.Destroy(t, terraformOptions)
-
-	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
-	terraform.InitAndApply(t, terraformOptions)
-}
-
-func TestDelegationSet(t *testing.T) {
-	t.Parallel()
-
-	terraformOptions := &terraform.Options{
-		// The path to where your Terraform code is located
-		TerraformDir: "../examples/delegation-set",
-		Upgrade:      true,
-	}
-
-	// At the end of the test, run `terraform destroy` to clean up any resources that were created
-	defer terraform.Destroy(t, terraformOptions)
-
-	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
-	terraform.InitAndApply(t, terraformOptions)
-}
-
-func TestPrivateHostedZone(t *testing.T) {
-	t.Parallel()
-
-	terraformOptions := &terraform.Options{
-		// The path to where your Terraform code is located
-		TerraformDir: "../examples/private-hosted-zone",
-		Upgrade:      true,
-	}
-
-	// At the end of the test, run `terraform destroy` to clean up any resources that were created
-	defer terraform.Destroy(t, terraformOptions)
-
-	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
-	terraform.InitAndApply(t, terraformOptions)
-}
-
-func TestMultipleDomainsSameRecords(t *testing.T) {
-	t.Parallel()
-
-	terraformOptions := &terraform.Options{
-		// The path to where your Terraform code is located
-		TerraformDir: "../examples/multiple-domains-same-records",
-		Upgrade:      true,
-	}
-
-	// At the end of the test, run `terraform destroy` to clean up any resources that were created
-	defer terraform.Destroy(t, terraformOptions)
-
-	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
-	terraform.InitAndApply(t, terraformOptions)
-}
-
-func TestMultipleDomainsDifferentRecords(t *testing.T) {
-	t.Parallel()
-
-	terraformOptions := &terraform.Options{
-		// The path to where your Terraform code is located
-		TerraformDir: "../examples/multiple-domains-different-records",
-		Upgrade:      true,
-	}
-
-	// At the end of the test, run `terraform destroy` to clean up any resources that were created
-	defer terraform.Destroy(t, terraformOptions)
-
-	// This will run `terraform init` and `terraform apply` and fail the test if there are any errors
-	terraform.InitAndApply(t, terraformOptions)
-}
